Store time properties internally as time_t
QDateTime is expensive to maintain and compare, so we can reduce
effort to build and maintain models by storing timestamp data in
time_t format, matching the database representation.

Because the API already provides QDateTime-based interfaces, we
maintain the time_t values in sync with original data, unless that
data has not been acessed; in this case it is null until referenced.